feat: save local and remote addr (#9)
This changeset implements saving the local and remote addr for all the
transports we support, improving data quality.

The changes should be obvious for all protocols but DoH, for which we
introduced a lower-level mechanism to collect this info, and namely the
HTTPClientDo function, opening up the possibility of advanced
customization code to create a connection on the fly. The previous
mechanism using HTTPClient is still working as intended.

Another minor change is refactoring the structured logging code inside
the slog.go and slog_test.go files.