BREAKING CHANGE: remove NewResolver NewTransport
This diff removes the NewResolver and NewTransport constructors,
which were just constructing the zero value.

Semantically, I expect a constructor to create a valid value and I was
not doing this. So, the simplest diff/fix for that is to just remove
the constructors and their implicit promise they are constructing.

Users of `dnscore` will need to update their code accordinggly:

- replace `NewTransport()` with `&Transport{}`;

- replace `NewResolver()` with `&Resolver{}`.